requirements.txt
bs4
lxml
nltk
torch
dill
pandas
apscheduler
sentence-transformers
requests
gensim==3.8.3
transformers
SentencePiece
newspaper3k
pyshorteners
selenium
git+https://github.com/PrithivirajDamodaran/Styleformer.git
